% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_count.R
\name{allele_count}
\alias{allele_count}
\alias{allele_count_from_genotype}
\title{Allele Count}
\usage{
allele_count(count, n)

allele_count_from_genotype(genotype, groups = NULL)
}
\arguments{
\item{count}{Matrix giving the number of alternative (or mutant) alleles
with variants as rows and samples as columns.}

\item{n}{Matrix giving the total number of alleles observed with variants as
rows and samples as columns.}

\item{genotype}{Sample genotype matrix with the variants as rows and samples
as column in the format \code{012}, where \code{0} is the reference homozygous, \code{1}
the heterozygous and \code{2} the alternative homozygous.}

\item{groups}{List of sample vector or a single vector of groups used to
compute the allele count. When is \emph{NULL} (default value) it considers all
samples as a single group, when is either a vector it considers only the
corresponding samples, and when is a list it computes the allele count for
each sub-vector.}
}
\value{
Allele count matrix
}
\description{
Creates an allele count matrix used on other analysis. The structure
encapsulates the allele count and number of individual with variant as rows
and groups as columns. It allows subsetting (\verb{[]}), to check dimensions
(with \code{dim}, \code{nrow}, and \code{ncol}), and renaming them.
}
\examples{
ac <- allele_count(matrix(0:5, ncol = 2), matrix(rep(6, 6), ncol = 2))

print(ac[1:2, ]) # allows subsetting

nrow(ac)         # number of variants
rownames(ac)     # variants identifier

ncol(ac)         # number of groups
colnames(ac)     # groups names
## From genotype matrix
geno_sample <- matrix(sample(0:2, 100, replace = TRUE), ncol = 10)
ac_matrix_from_genotype(geno_sample)      ## all samples
ac_matrix_from_genotype(geno_sample, 1:5) ## only the first 5
## for three groups
ac_matrix_from_genotype(
  geno_sample,
  list(a = 1:5, b = 3:7, c = 6:10))

}
