% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_frequency.R
\name{allele_frequency}
\alias{allele_frequency}
\title{The allele frequency from the allele count}
\usage{
allele_frequency(ac, rows = TRUE, cols = TRUE)
}
\arguments{
\item{ac}{A allele count matrix to compute allele frequency}

\item{rows, cols}{A vector indicating the subset of rows and columns to be
evaluated. If TRUE, no subseting is done.}
}
\description{
Given a allele count, it computes the allele frequency as numeric.
}
