% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_methods.R
\name{allele_methods}
\alias{count_allele}
\alias{total_allele}
\alias{invert_allele}
\alias{population_names}
\title{Allele Count Manipulation}
\usage{
count_allele(ac)

total_allele(ac)

invert_allele(ac)

population_names(ac)
}
\arguments{
\item{ac}{Allele count matrix to analyze}
}
\description{
Functions to explore and manipule the data from allele count matrices.
}
