% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_stat.R
\name{d_stat}
\alias{d_stat}
\title{Compute D-statistics}
\usage{
d_stat(ac, pops)
}
\arguments{
\item{ac}{A allele count matrix including the populations to investigate.}

\item{pops}{A vector of populations to compute the statistic.}
}
\description{
Computes the unbiased estimative of \emph{D-statististic}, also known as
4-population test, used to measure populations strutcture as described by
\href{http://www.genetics.org/content/192/3/1065}{Patterson et al. 2012}.
Giving 4 populations (\emph{W}, \emph{X}, \emph{Y}, and \emph{Z}), \emph{D-statistic} compares the
probability of a \emph{BABA} event, meaning \emph{W} and \emph{Y} agree and \emph{X} and \emph{Z}
agree, to the probability of a \emph{ABBA} event, meaning \emph{W} and \emph{Z} agree.
When \emph{D-statistic} present positive values it favors \emph{BABA}, when negative
present \emph{ABBA}, and zero favors \emph{AABB} (\emph{W} and \emph{X} agree). The
\emph{D-statistic} describe the following equation:
\deqn{D = (P(BABA) - P(ABBA)) / (P(BABA) + P(ABBA))}
}
