% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat}
\alias{f_stat}
\alias{f4_stat}
\alias{f3_stat}
\alias{f2_stat}
\title{Compute Reich's f-statistics}
\usage{
f4_stat(ac, pops)

f3_stat(ac, pops)

f2_stat(ac, pops)
}
\arguments{
\item{ac}{A allele count matrix including the populations to investigate.}

\item{pops}{A vector of populations to compute the statistic.}
}
\description{
Computes \emph{f2}, \emph{f3}, and \emph{f4} statistics unbiased estimateve used to measure
populations strutcture and treeness as described by
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2842210/}{Reich et al. 2009}.
The \emph{f-statistic} are described the following statistics:
\deqn{F4(W, X; Y,Z) = E[(W - X) * (Y - Z)]}
\deqn{F3(W; X, Y) = E[(W - X) * (W - Y)]}
\deqn{F2(X, Y) = E[(X - Y)^2]}
}
