% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variants.R
\name{filter_variants}
\alias{filter_variants}
\alias{complete_variants}
\alias{informative_variants}
\title{Filter Variants from Allele Count}
\usage{
filter_variants(x, fn, cols = NULL)

complete_variants(x, cols = NULL)

informative_variants(x, cols = NULL)
}
\arguments{
\item{x}{A allele count or frequency matrix to be filtered.}

\item{cols}{A vector indicating the subset of columns considered.}
}
\description{
Implements common strategies to filter variants of an allele frequency
matrix, including \code{complete_variants}, which exclude variants with missing
genotype, and \code{informative_variants}, which excludes variants were allgroups
are reference or alternative homozygous and keeping only informative sites.
These functions return a \code{boolean} vector of the subset and can also be used
with \code{filter_variants} to return the subsetted allele count or frequency
matrix.
}
