% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\docType{data}
\name{fst}
\alias{fst_methods}
\alias{fst}
\alias{bootstrap_fst}
\title{Calculate the pairwise Fst statistic}
\format{An object of class \code{character} of length 5.}
\usage{
fst_methods

fst(ac, method = fst_methods, snpwise = FALSE)

bootstrap_fst(ac, method = fst_methods, boots = 100L, block = 10L,
  as_matrix = FALSE)
}
\arguments{
\item{ac}{allele count matrix for which to compute \code{Fst}.}

\item{method}{fst estimator method. Using \code{reich} as default.}

\item{snpwise}{a boolean scalar idicating if should return a the statistic
for each variant (when \code{TRUE}) or the summary (when \code{FALSE}).}

\item{boots}{number of bootstrap replicates to perform}

\item{block}{statistic blocking size}

\item{as_matrix}{if \code{TRUE}, then return a matrix whith bootstrap interations
as rows and pairwise comparisons as columns.}
}
\description{
Given a \code{allele_count} matrix, it computes the \emph{Fst} statistic and its
bootstrap replications between all pairs of population using several
methods.
}
\details{
The estimator methods include \emph{Reich et al. (2019)} (\code{reich}),
\emph{Weir & Cockerham (1984)} (\code{weir_cockerham} or \code{wc}), \emph{Hudson et al. (1992)}
(\code{hudson}) and \emph{Wright (1951)} (\code{wright}). These methods use different
approaches to estimate \emph{Fst}. The \emph{Wright} method is a biased estimator and
can be affected by different populations sizes, \emph{Hudson} and \emph{Reich} methods
are unbiased fast and tend to present reliable results on large variant
sizes, but limited precision snpwise. In other hand, \emph{Weir & Cockerham}
method is the most common estimator (implemented by tools such as \emph{plink},
\emph{vcftools}, and \emph{scikit-allele}) and uses a population size correction to
reduce bias, but is a little slower than the other methods.

Please note that \emph{Reich} method requires population sizes larger than 2 (
more than one diploid individual).
}
\examples{
c <- matrix(sample(1:10, 250, replace = TRUE), ncol = 10, nrow = 25)
n <- matrix(10, ncol = 10, nrow = 25)
ac <- allele_count(c, n)
## Summarized Fst
fst(ac)
fst(ac, "wc")
## Snpwise Fst
fst(ac, snpwise = TRUE)
fst(ac, "wc", snpwise = TRUE)

## bootstrap fst
bootstrap_fst(ac)

}
\keyword{datasets}
