\name{htspop-package}
\alias{htspop-package}
\alias{htspop}
\docType{package}
\title{
\packageTitle{htspop}
}
\description{
\packageDescription{htspop}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{htspop}
\packageIndices{htspop}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{htspop}

Maintainer: \packageMaintainer{htspop}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
