% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{A blocked jacknife estimator}
\usage{
jackknife(x, block = 10, weighted = TRUE)
}
\arguments{
\item{x}{A numeric vector or matrix to estimate mean and standard error.}

\item{block}{A integer scalar indicating the size of blocks (greater
than zero).}

\item{weighted}{A boolean scalar indicating if should perform a weighted
estimation.}
}
\description{
It performs a blocked jacknife estimator of mean and standard error of the
input vector. Given a matrix as input, it considers the first column as
numerator and second column as denominator. It can also perform a weighted
estimation by counting the number of non-NA values per block.
\emph{This implementation is largely inspired on
(Treemix)\link{https://bitbucket.org/nygcresearch/treemix/wiki/Home}
implementation of the algorithm}.
}
