% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nei.R
\docType{data}
\name{nei}
\alias{nei_methods}
\alias{nei}
\alias{bootstrap_nei}
\title{Calculate the pairwise Nei distance}
\format{
An object of class \code{character} of length 3.
}
\usage{
nei_methods

nei(ac, method = nei_methods)

bootstrap_nei(
  ac,
  method = nei_methods,
  boots = 100L,
  block = 10L,
  as_matrix = FALSE
)
}
\arguments{
\item{ac}{allele count matrix for which to compute nei's distance.}

\item{method}{nei's distance estimator method, with \code{fast} as default.}

\item{boots}{number of bootstrap replicates to perform}

\item{block}{statistic blocking size}

\item{as_matrix}{if \code{TRUE}, then return a matrix whith bootstrap interations
as rows and pairwise comparisons as columns.}
}
\description{
Given a \code{allele_count} matrix, it computes \emph{Nei} distance and its bootstrap
replications between all pairs of population using different methods.
}
\details{
The estimator methods include two implementations of the \emph{standard Nei's}
genetic distance (\code{fast} and \code{safe}), as described in \emph{Nei (1972)}, and
\emph{Nei's Da} (\code{da}), described in \emph{Nei et al. (1983)}. Both Nei's distance
(\emph{standard} and \emph{Da}) measures genetic differences caused by mutation and
genetic drift with \emph{Da} result in more reliable results for microsatellite
data, but more sensible to small population sizes.

Two versions of \emph{standard Nei's distance} were implemented, \code{safe} and
\code{fast}. The \code{fast} implementation is 2 times faster than \code{safe}, but can
introduce errors when there is data missing, if no missing data is present
it results the same result as the \code{safe} implementation.
}
\examples{
c <- matrix(sample(1:10, 250, replace = TRUE), ncol = 10, nrow = 25)
n <- matrix(10, ncol = 10, nrow = 25)
ac <- allele_count(c, n)
## Summarized Nei
nei(ac)
nei(ac, "safe")

bootstrap_nei(ac)

}
\keyword{datasets}
